# bundles/de.cooperateproject.cdo.dawn.frontend/web_content/swagger-fixed.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Dawn Web API",
    "description": "The Dawn Web API provides access to a dawn repository and utility methods."
  },
  "host": "localhost:9090",
  "basePath": "/services",
  "tags": [
    {
      "name": "browse"
    }
  ],
  "paths": {
    "/browse": {
      "get": {
        "tags": [
          "browse"
        ],
        "summary": "Lists all projects",
        "description": "",
        "operationId": "getProjects",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Project"
              }
            }
          }
        }
      }
    },
    "/browse/{projectid}": {
      "get": {
        "tags": [
          "browse"
        ],
        "summary": "Get project info",
        "description": "",
        "operationId": "getProject",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "projectid",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Project"
            }
          }
        }
      }
    },
    "/browse/{projectid}/models": {
      "get": {
        "tags": [
          "browse"
        ],
        "summary": "Lists all models",
        "description": "",
        "operationId": "getModels",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "projectid",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Model"
              }
            }
          }
        }
      }
    },
    "/browse/{projectid}/models/{modelid}": {
      "get": {
        "tags": [
          "browse"
        ],
        "summary": "Get model info",
        "description": "",
        "operationId": "getModel",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "projectid",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "modelid",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Model"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Project": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "models": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Model"
          }
        }
      }
    },
    "Model": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        }
      }
    }
  }
}